# api/openapi.yaml
openapi: 3.1.0
info:
  title: sing-box-panel API
  version: 0.1.0
  license:
    name: GPL-3.0-or-later
    identifier: GPL-3.0-or-later
servers:
  - url: /api/v1
paths:
...
  /logs:
    get:
      operationId: listDurableLogs
      description: Lists sanitized durable event metadata newest first. It never returns full configuration, subscription content, or credentials.
      parameters:
        - name: source
          in: query
          schema: { $ref: "#/components/schemas/LogSource" }
        - name: level
          in: query
          schema: { $ref: "#/components/schemas/LogLevel" }
        - name: code
          in: query
          schema: { type: string, pattern: "^[a-z][a-z0-9_.-]*$" }
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: until
          in: query
          schema: { type: string, format: date-time }
        - name: after_time
          in: query
          description: Exclusive newest-first cursor time; requires after_id.
          schema: { type: string, format: date-time }
        - name: after_id
          in: query
          description: Exclusive newest-first cursor ID; requires after_time.
          schema: { type: string }
        - $ref: "#/components/parameters/PageLimit"
      responses:
        "200":
          description: Sanitized durable log page.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogPage" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: clearDurableLogs
      description: Explicitly deletes sanitized durable logs matching the optional source and exclusive before cutoff.
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - name: source
          in: query
          schema: { $ref: "#/components/schemas/LogSource" }
        - name: before
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Number of deleted entries.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogClearResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
  /logs/stream:
    get:
      operationId: streamDurableLogs
      description: Streams sanitized durable log entries as SSE. Without since, an after cursor, or Last-Event-ID, the stream begins at connection time. Event IDs encode the durable time/id cursor for reconnects.
      parameters:
        - name: source
          in: query
          schema: { $ref: "#/components/schemas/LogSource" }
        - name: level
          in: query
          schema: { $ref: "#/components/schemas/LogLevel" }
        - name: code
          in: query
          schema: { type: string, pattern: "^[a-z][a-z0-9_.-]*$" }
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: after_time
          in: query
          description: Exclusive cursor time; requires after_id and cannot be combined with since.
          schema: { type: string, format: date-time }
        - name: after_id
          in: query
          description: Exclusive cursor ID; requires after_time and cannot be combined with since.
          schema: { type: string }
        - $ref: "#/components/parameters/PageLimit"
      responses:
        "200":
          description: A text/event-stream of log events with keepalive comments.
          content:
            text/event-stream:
              schema: { type: string }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
  /logs/{logId}:
    parameters:
      - name: logId
        in: path
        required: true
        schema: { type: string, minLength: 1, maxLength: 80 }
    get:
      operationId: getDurableLog
      responses:
        "200":
          description: One sanitized durable event.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogEntry" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteDurableLog
      description: Explicitly deletes one sanitized durable log entry.
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "200":
          description: Stable deletion acknowledgement.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogDeleteResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
...
  /traffic/status:
    get:
      operationId: getTrafficStatus
      description: Returns the same applied-bundle evidence contract as current metrics.
      responses:
        "200":
          description: Current traffic collector availability and data.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MetricsSnapshot" }
        "401": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /traffic/periods:
    get:
      operationId: listTrafficPeriods
      parameters:
        - name: activation_bundle_id
          in: query
          schema: { type: string }
        - name: from
          in: query
          description: Include periods whose end is later than this instant.
          schema: { type: string, format: date-time }
        - name: to
          in: query
          description: Include periods whose start is earlier than this instant.
          schema: { type: string, format: date-time }
        - $ref: "#/components/parameters/PageLimit"
      responses:
        "200":
          description: Persisted collector periods, newest first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrafficPeriodPage" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
  /traffic/periods/{periodId}:
    parameters:
      - name: periodId
        in: path
        required: true
        schema: { type: string, minLength: 1, maxLength: 256 }
    get:
      operationId: getTrafficPeriod
      responses:
        "200":
          description: One persisted collector period.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrafficPeriod" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    managementBearer:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: sbp_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required for cookie-authenticated state-changing requests.
      schema:
        type: string
        minLength: 1
...
    PageLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    Problem:
      description: Stable RFC 9457-compatible problem response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        request_id: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
...
    MonitoringTier:
      type: string
      enum: [full, limited, process_only]
...
    LogSource:
      type: string
      enum: [panel, core, task, security]
    LogLevel:
      type: string
      enum: [trace, debug, info, warn, error, fatal]
    LogEntry:
      type: object
      additionalProperties: false
      required: [id, time, source, level, code, message, metadata]
      properties:
        id: { type: string }
        time: { type: string, format: date-time }
        source: { $ref: "#/components/schemas/LogSource" }
        level: { $ref: "#/components/schemas/LogLevel" }
        code: { type: string }
        message: { type: string }
        metadata:
          type: object
          additionalProperties: true
          description: Sanitized bounded metadata; credentials and complete configuration/subscription bodies are never persisted.
    LogCursor:
      type: object
      additionalProperties: false
      required: [time, id]
      properties:
        time: { type: string, format: date-time }
        id: { type: string }
    LogPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/LogEntry" }
        next: { $ref: "#/components/schemas/LogCursor" }
    LogClearResult:
      type: object
      additionalProperties: false
      required: [deleted]
      properties:
        deleted: { type: integer, format: int64, minimum: 0 }
    LogDeleteResult:
      type: object
      additionalProperties: false
      required: [id, deleted]
      properties:
        id: { type: string }
        deleted: { type: boolean, const: true }
    TrafficPeriod:
      type: object
      additionalProperties: false
      required: [id, period_start, period_end, inbound_bytes, outbound_bytes, counters, created_at]
      properties:
        id: { type: string }
        activation_bundle_id: { type: string }
        period_start: { type: string, format: date-time }
        period_end: { type: string, format: date-time }
        inbound_bytes: { type: integer, format: int64, minimum: 0 }
        outbound_bytes: { type: integer, format: int64, minimum: 0 }
        counters: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
    TrafficPeriodPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TrafficPeriod" }
    MetricsSnapshot:
      type: object
      additionalProperties: false
      required: [available, collected_at]
      properties:
        available: { type: boolean }
        reason_code:
          type: string
          enum: [not_applied, process_only, no_collector_sample, stale_collector_sample]
        applied_bundle_id: { type: string }
        monitoring_tier: { $ref: "#/components/schemas/MonitoringTier" }
        collected_at: { type: string, format: date-time }
        current_traffic_period: { $ref: "#/components/schemas/TrafficPeriod" }
...
security:
  - sessionCookie: []
  - managementBearer: []